tlshd: add 'keyring' configuration option
Looking up keys is only possible if the 'read' permission has been
set. As we do not want anyone to read the contents of a key we
need to ensure that the handshake daemon has the appropriate
permissions to do so.
This patch adds an 'keyring' configuration option to link a keyring
into the default 'session' keyring for the tlshd daemon; that way the
daemon always has possession of the requested keys and will
be able to read them.

Signed-off-by: Hannes Reinecke <hare@suse.de>