tls13: do not use callback to set client identity
The client identity is known prior to start the handshake (as we
only have one identity to cope with), so we should set it before
starting the handshake and avoid using the callback.
This has the benefit that we don't need to a global variable holding
the identity, which will be racy for several concurrent handshakes.

Signed-off-by: Hannes Reinecke <hare@suse.de>